+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Some performance improvements were done to `CTRFileIO` and `RomFSReader`.
 * Update copyright year
 * Various documentation and type hint changes
   * Add new `FilePath` and `FilePathOrObject` types
+* Add `__slots__` to `SubsectionIO` and `SplitFileMerger`
 
 ## v0.6.0 - January 26, 2022
 ### Highlights